split AzureStorage into dedicated nuget packages
