+++ .github/workflows/healthchecks_azurestorage_cd.yml
@@ -23,7 +23,19 @@ jobs:
         run: dotnet restore ./src/HealthChecks.AzureStorage/HealthChecks.AzureStorage.csproj
       - name: Build
         run: dotnet build --no-restore ./src/HealthChecks.AzureStorage/HealthChecks.AzureStorage.csproj -c $BUILD_CONFIG
+      - name: Pack Blobs
+        run: dotnet pack --no-build ./src/HealthChecks.Azure.Storage.Blobs/HealthChecks.Azure.Storage.Blobs.csproj -c $BUILD_CONFIG -o ./artifacts
+      - name: Pack Queues
+        run: dotnet pack --no-build ./src/HealthChecks.Azure.Storage.Queues/HealthChecks.Azure.Storage.Queues.csproj -c $BUILD_CONFIG -o ./artifacts
+      - name: Pack Files.Shares
+        run: dotnet pack --no-build ./src/HealthChecks.Azure.Storage.Files.Shares/HealthChecks.Azure.Storage.Files.Shares.csproj -c $BUILD_CONFIG -o ./artifacts
       - name: Pack
         run: dotnet pack --no-build ./src/HealthChecks.AzureStorage/HealthChecks.AzureStorage.csproj -c $BUILD_CONFIG -o ./artifacts
+      - name: Publish Blobs
+        run: dotnet nuget push ./artifacts/AspNetCore.HealthChecks.Azure.Storage.Blobs*.nupkg -k ${{secrets.NUGET_API_KEY}} -s https://api.nuget.org/v3/index.json --skip-duplicate
+      - name: Publish Queues
+        run: dotnet nuget push ./artifacts/AspNetCore.HealthChecks.Azure.Storage.Queues*.nupkg -k ${{secrets.NUGET_API_KEY}} -s https://api.nuget.org/v3/index.json --skip-duplicate
+      - name: Publish Files.Shares
+        run: dotnet nuget push ./artifacts/AspNetCore.HealthChecks.Azure.Storage.Files.Shares*.nupkg -k ${{secrets.NUGET_API_KEY}} -s https://api.nuget.org/v3/index.json --skip-duplicate
       - name: Publish
         run: dotnet nuget push ./artifacts/AspNetCore.HealthChecks.AzureStorage.*.nupkg -k ${{secrets.NUGET_API_KEY}} -s https://api.nuget.org/v3/index.json --skip-duplicate
