+++ .github/workflows/healthchecks_azurestorage_ci.yml
@@ -5,6 +5,7 @@ on:
   push:
    branches: [ master ]
    paths:
+     - src/HealthChecks.Azure.Storage.*/**
      - src/HealthChecks.AzureStorage/**
      - test/HealthChecks.AzureStorage.Tests/**
      - test/_SHARED/**
@@ -18,6 +19,7 @@ on:
   pull_request:
     branches: [ master ]
     paths:
+      - src/HealthChecks.Azure.Storage.*/**
       - src/HealthChecks.AzureStorage/**
       - test/HealthChecks.AzureStorage.Tests/**
       - test/_SHARED/**
