+++ src/HealthChecks.Azure.Storage.Blobs/HealthChecks.Azure.Storage.Blobs.csproj
@@ -1,0 +1,19 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+    <PackageTags>$(PackageTags);Azure;AzureStorage</PackageTags>
+    <Description>HealthChecks.Azure.Storage.Blobs is the health check package for Blobs.</Description>
+    <VersionPrefix>$(HealthCheckAzureStorage)</VersionPrefix>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <Compile Include="../ClientCache.cs" Link="ClientCache.cs" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Azure.Storage.Blobs" Version="12.17.0" />
+    <PackageReference Include="Microsoft.Extensions.Diagnostics.HealthChecks" Version="7.0.9" />
+  </ItemGroup>
+
+</Project>
