+++ src/HealthChecks.Azure.Storage.Files.Shares/AzureFileShareHealthCheck.cs
