+++ src/HealthChecks.Azure.Storage.Files.Shares/AzureFileShareHealthCheckOptions.cs
