+++ src/HealthChecks.AzureStorage/HealthChecks.AzureStorage.csproj
@@ -3,15 +3,14 @@
   <PropertyGroup>
     <TargetFramework>netstandard2.0</TargetFramework>
     <PackageTags>$(PackageTags);Azure;AzureStorage</PackageTags>
-    <Description>HealthChecks.AzureStorage is the health check package for Blobs, Tables and Queues.</Description>
+    <Description>HealthChecks.AzureStorage is the health check package for Blobs, Tables and Files Shares.</Description>
     <VersionPrefix>$(HealthCheckAzureStorage)</VersionPrefix>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Azure.Storage.Blobs" Version="12.17.0" />
-    <PackageReference Include="Azure.Storage.Queues" Version="12.15.0" />
-    <PackageReference Include="Azure.Storage.Files.Shares" Version="12.15.0" />
-    <PackageReference Include="Microsoft.Extensions.Diagnostics.HealthChecks" Version="7.0.9" />
+    <ProjectReference Include="..\HealthChecks.Azure.Storage.Blobs\HealthChecks.Azure.Storage.Blobs.csproj" />
+    <ProjectReference Include="..\HealthChecks.Azure.Storage.Files.Shares\HealthChecks.Azure.Storage.Files.Shares.csproj" />
+    <ProjectReference Include="..\HealthChecks.Azure.Storage.Queues\HealthChecks.Azure.Storage.Queues.csproj" />
   </ItemGroup>
 
 </Project>
