+++ src/HealthChecks.AzureStorage/Properties/AssemblyInfo.cs
@@ -1,0 +1,9 @@
+using System.Runtime.CompilerServices;
+using HealthChecks.AzureStorage;
+
+[assembly: TypeForwardedTo(typeof(AzureBlobStorageHealthCheck))]
+[assembly: TypeForwardedTo(typeof(AzureBlobStorageHealthCheckOptions))]
+[assembly: TypeForwardedTo(typeof(AzureQueueStorageHealthCheck))]
+[assembly: TypeForwardedTo(typeof(AzureQueueStorageHealthCheckOptions))]
+[assembly: TypeForwardedTo(typeof(AzureFileShareHealthCheck))]
+[assembly: TypeForwardedTo(typeof(AzureFileShareHealthCheckOptions))]
